feat: 优化 OLED 模块功能和代码结构
- 重构了 OLED_ShowChar、OLED_ShowNum、OLED_ShowChinese 等函数，提高了代码可读性和可维护性
- 新增 OLED_ShowString 和 OLED_Printf 函数，支持更灵活的字符串显示
- 新增 OLED_Data.h 文件，用于管理字模和图像数据
- 优化了代码注释和命名规范，提高了代码的可读性

+++ src/System/PWM.c
@@ -45,7 +45,7 @@ void PWM_Disable(void)
     TIM_CtrlPWMOutputs(TIM1, DISABLE);
 }
 
-void PWM_Set(uint8_t Percent)
+void PWM_Set(uint16_t Percent)
 {
     if (Percent > 100)
         Percent = 100;
